Partner preview block (#58)
Adds a partner preview block (front page partner list with images + partner "level").

I added a `Partner type` to the configuration where you can select what partner types are available, along with referencing this for each `Partner`.

I also added a Typescript model for the `Partner` document since it was missing.

Should probably be its own PR but whatever: also moved out the "sub-heading" into its own component (the heading with a pink line next to it), since I had to re-use it.

The styling is very much less than optimal and more of a quick draft right now. I'm not sure it makes much sense to spend hours on making things pixel perfect until more of the web site is fleshed out and it can be seen in a proper context.